requirements.txt
aiohttp==3.3.2
aiohttp_cors==0.7.0
alembic==0.9.10
aiopg==0.14.0
SQLAlchemy==1.2.9
psycopg2-binary==2.7.5
alembic==0.9.10
gunicorn==19.9.0
graphene==2.1.8
argparse==1.4.0
aiohttp-graphql==1.0.0
trafaret-config==2.0.2
trafaret==1.2.0
graphql-core==2.3.1